Rename ui/userprofile package to ui/user
