+++ app/src/main/java/org/ole/planet/myplanet/ui/user/ProfileAdapter.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.userprofile
+package org.ole.planet.myplanet.ui.user
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
