+++ app/src/main/java/org/ole/planet/myplanet/ui/user/UserProfileViewModel.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.userprofile
+package org.ole.planet.myplanet.ui.user
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
